[muxorch] run updateroutes before disbling neighbor during update
fdb update from an active mux port to a standby mux port when a static
route was present caused orchagent crash. This is because the route was
still pointing to the active neighbor nexthop when disabling this
neighbor. This calls updateNexthopRoutes before disabling standby
neighbor (but after neighbor nh is updated in muxorch) to set the route
nexthop to the tunnel route.

Signed-off-by: Nikola Dancejic <ndancejic@microsoft.com>